refactor: the password manager package ui & feature based package structure
