+++ app/src/main/kotlin/com/jeeldobariya/passcodes/navigation/NavigationRoot.kt
@@ -14,14 +14,14 @@ import androidx.navigation3.ui.NavDisplay
 import com.jeeldobariya.passcodes.core.datastore.AppSettings
 import com.jeeldobariya.passcodes.core.datastore.appDatastore
 import com.jeeldobariya.passcodes.core.navigation.Route
-import com.jeeldobariya.passcodes.password_manager.ui.ClassicalLoadPasswordScreen
-import com.jeeldobariya.passcodes.password_manager.ui.ClassicalPasswordManagerScreen
-import com.jeeldobariya.passcodes.password_manager.ui.ClassicalSavePasswordScreen
-import com.jeeldobariya.passcodes.password_manager.ui.ClassicalUpdatePasswordScreen
-import com.jeeldobariya.passcodes.password_manager.ui.ClassicalViewPasswordScreen
-import com.jeeldobariya.passcodes.password_manager.ui.ModernPasswordManagerScreen
-import com.jeeldobariya.passcodes.password_manager.ui.ModernSavePasswordScreen
-import com.jeeldobariya.passcodes.password_manager.ui.ModernUpdatePasswordScreen
+import com.jeeldobariya.passcodes.password_manager.presentation.load_password.ClassicalLoadPasswordScreen
+import com.jeeldobariya.passcodes.password_manager.presentation.password_manager.ClassicalPasswordManagerScreen
+import com.jeeldobariya.passcodes.password_manager.presentation.password_manager.ModernPasswordManagerScreen
+import com.jeeldobariya.passcodes.password_manager.presentation.save_password.ClassicalSavePasswordScreen
+import com.jeeldobariya.passcodes.password_manager.presentation.save_password.ModernSavePasswordScreen
+import com.jeeldobariya.passcodes.password_manager.presentation.update_password.ClassicalUpdatePasswordScreen
+import com.jeeldobariya.passcodes.password_manager.presentation.update_password.ModernUpdatePasswordScreen
+import com.jeeldobariya.passcodes.password_manager.presentation.view_password.ClassicalViewPasswordScreen
 import com.jeeldobariya.passcodes.ui.ClassicalAboutScreen
 import com.jeeldobariya.passcodes.ui.ClassicalMainScreen
 import com.jeeldobariya.passcodes.ui.ClassicalSettingsScreen
