+++ password_manager/src/main/kotlin/com/jeeldobariya/passcodes/password_manager/presentation/password_manager/PasswordManagerAction.kt
@@ -1,0 +1,5 @@
+package com.jeeldobariya.passcodes.password_manager.presentation.password_manager
+
+sealed interface PasswordManagerAction {
+    data object RefreshPassword : PasswordManagerAction
+}
