+++ password_manager/src/main/kotlin/com/jeeldobariya/passcodes/password_manager/presentation/save_password/ModernSavePasswordScreen.kt
@@ -1,4 +1,4 @@
-package com.jeeldobariya.passcodes.password_manager.ui
+package com.jeeldobariya.passcodes.password_manager.presentation.save_password
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -19,11 +19,11 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.jeeldobariya.passcodes.password_manager.presentation.save_password.SavePasswordAction
-import com.jeeldobariya.passcodes.password_manager.presentation.save_password.SavePasswordViewModel
 import kotlinx.coroutines.launch
 import org.koin.compose.viewmodel.koinViewModel
+
 
 @Composable
 fun ModernSavePasswordScreen(viewmodel: SavePasswordViewModel = koinViewModel()) {
@@ -101,88 +101,8 @@ fun ModernSavePasswordScreen(viewmodel: SavePasswordViewModel = koinViewModel())
     }
 }
 
-
+@Preview
 @Composable
-fun ClassicalSavePasswordScreen(viewmodel: SavePasswordViewModel = koinViewModel()) {
-    val snackbarHostState = remember { SnackbarHostState() }
-    val scope = rememberCoroutineScope()
-
-    Scaffold(
-        snackbarHost = { SnackbarHost(hostState = snackbarHostState) },
-    ) { paddingValues ->
-        val state by viewmodel.state.collectAsState()
-
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(paddingValues),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text("Save Password")
-
-            Spacer(modifier = Modifier.padding(16.dp))
-
-            OutlinedTextField(
-                value = state.domain,
-                onValueChange = {
-                    viewmodel.onAction(action = SavePasswordAction.OnChangeDomain(it))
-                },
-                label = {
-                    Text("Domain:")
-                },
-                placeholder = {
-                    Text("e.g. Google, Instagram etc...")
-                }
-            )
-
-            OutlinedTextField(
-                value = state.username,
-                onValueChange = {
-                    viewmodel.onAction(action = SavePasswordAction.OnChangeUsername(it))
-                },
-                label = {
-                    Text("Username:")
-                },
-                placeholder = {
-                    Text("e.g. username / email / mobile.no.")
-                }
-            )
-
-            OutlinedTextField(
-                value = state.password,
-                onValueChange = {
-                    viewmodel.onAction(action = SavePasswordAction.OnChangePassword(it))
-                },
-                label = {
-                    Text("Password:")
-                },
-                visualTransformation = PasswordVisualTransformation()
-            )
-
-            OutlinedTextField(
-                value = state.notes,
-                onValueChange = {
-                    viewmodel.onAction(action = SavePasswordAction.OnChangeNotes(it))
-                },
-                label = {
-                    Text("Notes (Optional):")
-                },
-                placeholder = {
-                    Text("e.g. Url or Platform Info. Account Info.")
-                }
-            )
-
-            Spacer(modifier = Modifier.padding(8.dp))
-
-            Button(onClick = {
-                viewmodel.onAction(action = SavePasswordAction.OnSavePasswordButtonClick)
-                scope.launch {
-                    snackbarHostState.showSnackbar("Saved Successfully!!")
-                }
-            }) {
-                Text("Save Password")
-            }
-        }
-    }
+fun ModernSavePasswordScreenPreview() {
+    ModernSavePasswordScreen()
 }
