+++ password_manager/src/main/kotlin/com/jeeldobariya/passcodes/password_manager/presentation/view_password/ClassicalViewPasswordScreen.kt
@@ -1,4 +1,4 @@
-package com.jeeldobariya.passcodes.password_manager.ui
+package com.jeeldobariya.passcodes.password_manager.presentation.view_password
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -13,11 +13,10 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.jeeldobariya.passcodes.core.navigation.Route
-import com.jeeldobariya.passcodes.password_manager.presentation.view_password.ViewPasswordAction
-import com.jeeldobariya.passcodes.password_manager.presentation.view_password.ViewPasswordViewModel
 import org.koin.compose.viewmodel.koinViewModel
 
 @Composable
@@ -79,4 +78,10 @@ fun ClassicalViewPasswordScreen(
             }
         }
     }
+}
+
+@Preview
+@Composable
+fun ClassicalViewPasswordScreenPreview() {
+    ClassicalViewPasswordScreen(passwordId = 0, navigateTo = {})
 }
